Split lnd client into separate module
